Move array functions to lib
