+++ dbms/src/Functions/array/arrayEnumerateExtended.h
