+++ dbms/src/Functions/array/arrayExists.cpp
@@ -1,6 +1,6 @@
 #include <DataTypes/DataTypesNumber.h>
 #include <Columns/ColumnsNumber.h>
-#include <Functions/FunctionArrayMapped.h>
+#include "FunctionArrayMapped.h"
 #include <Functions/FunctionFactory.h>
