+++ dbms/src/Functions/array/countEqual.cpp
@@ -1,4 +1,4 @@
-#include <Functions/arrayIndex.h>
+#include "arrayIndex.h"
 #include <Functions/FunctionFactory.h>
