+++ dbms/src/Functions/array/emptyArrayToSingle.cpp
